Split Radarr models into dedicated files
