+++ Cantinarr/Features/Radarr/Models/RadarrQualityProfile.swift
@@ -1,0 +1,9 @@
+// File: RadarrQualityProfile.swift
+// Purpose: Defines RadarrQualityProfile component for Cantinarr
+
+import Foundation
+
+struct RadarrQualityProfile: Codable, Identifiable {
+    let id: Int
+    let name: String
+}
