+++ Cantinarr/Features/Radarr/Models/RadarrSystemStatus.swift
@@ -1,0 +1,10 @@
+// File: RadarrSystemStatus.swift
+// Purpose: Defines RadarrSystemStatus component for Cantinarr
+
+import Foundation
+
+struct RadarrSystemStatus: Codable {
+    let appName: String?
+    let version: String
+    let buildTime: Date?
+}
