+++ Cantinarr/Features/Radarr/Models/RadarrUnmappedFolder.swift
@@ -1,0 +1,9 @@
+// File: RadarrUnmappedFolder.swift
+// Purpose: Defines RadarrUnmappedFolder component for Cantinarr
+
+import Foundation
+
+struct RadarrUnmappedFolder: Codable, Hashable, Equatable {
+    let name: String?
+    let path: String?
+}
